+++ packages/core/src/DropDownMenu/DropDownMenu.styles.tsx
@@ -1,7 +1,7 @@
 import { createClasses } from "@hitachivantara/uikit-react-utils";
 import { theme } from "@hitachivantara/uikit-styles";
 
-export const { staticClasses, useClasses } = createClasses("HvDropDownMenu", {
+export const { staticClasses, useClasses } = createClasses("HvDropdownMenu", {
   root: {
     flexShrink: 0,
     "& > div": {